refactor: move generate subscription object logic to library
