+++ src/components/AppShell/SettingsModal.tsx
@@ -6,7 +6,7 @@ import { Modal, Button, Checkbox } from "@mantine/core";
 import { useQuery, useMutation } from "@tanstack/react-query";
 import { getUserSettings } from "@/lib/requests/settings/getUserSettings";
 import { postSubscribe } from "@/lib/requests/settings/postSubscribe";
-import type { WebPushSubscription } from "@/types/users.types";
+import { generateSubscriptionObject } from "@/lib/requests/settings/postSubscribe";
 import { putUserSettings } from "@/lib/requests/settings/putUserSettings";
 
 type SettingModalProps = {
@@ -37,32 +37,6 @@ export default function SettingsModal({
             setNewReferenceToMyReply(data.notification.newReferenceToMyReply);
         }
     }, [data]);
-
-    const arrayBufferToBase64 = (buffer: ArrayBuffer | null): string => {
-        if (!buffer) return "";
-        return btoa(String.fromCharCode(...new Uint8Array(buffer)));
-    };
-
-    const generateSubscriptionObject =
-        async (): Promise<WebPushSubscription> => {
-            // Get a PushSubscription object
-            const serviceWorker = new ServiceWorkerRegistration();
-            const pushSubscription =
-                await serviceWorker.pushManager.subscribe();
-
-            // Create an object containing the information needed by the app server
-            const subscriptionObject = {
-                endpoint: pushSubscription.endpoint,
-                keys: {
-                    p256dh: arrayBufferToBase64(
-                        pushSubscription.getKey("p256dh"),
-                    ),
-                    auth: arrayBufferToBase64(pushSubscription.getKey("auth")),
-                },
-            };
-
-            return subscriptionObject;
-        };
 
     const postSubscribeMutation = useMutation({
         mutationKey: ["postSubscribe"],
